requirements.txt
pandas
pandas_gbq
sklearn
numpy 
keras
tensorflow==1.11.0 
jsonpickle 
google-cloud 
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery

